Move factories to namespace
