+++ src/SpiffyAuthorize/Factory/RbacServiceFactory.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace SpiffyAuthorize\Service;
+namespace SpiffyAuthorize\Factory;
 
+use SpiffyAuthorize\Service\RbacService;
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
