+++ src/SpiffyAuthorize/Factory/ViewStrategyUnauthorizedFactory.php
@@ -1,26 +1,26 @@
 <?php
 
-namespace SpiffyAuthorize\Service;
+namespace SpiffyAuthorize\Factory;
 
-use SpiffyAuthorize\Collector\RoleCollector;
-use SpiffyAuthorize\Guard\RouteGuard;
+use SpiffyAuthorize\View\UnauthorizedStrategy;
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
 
-class CollectorRoleFactory implements FactoryInterface
+class ViewStrategyUnauthorizedFactory implements FactoryInterface
 {
     /**
      * Create service
      *
      * @param ServiceLocatorInterface $serviceLocator
-     * @return RoleCollector
+     * @return mixed
      */
     public function createService(ServiceLocatorInterface $serviceLocator)
     {
         /** @var \SpiffyAuthorize\ModuleOptions $options */
         $options  = $serviceLocator->get('SpiffyAuthorize\ModuleOptions');
-        $provider = $serviceLocator->get($options->getIdentityProvider());
+        $strategy = new UnauthorizedStrategy();
+        $strategy->setTemplate($options->getViewTemplate());
 
-        return new RoleCollector($provider);
+        return $strategy;
     }
 }
